+++ src/internal/dto/list_queries.go
@@ -15,8 +15,8 @@ type MouListQuery struct {
 
 type ClientListQuery struct {
 	PaginationQuery
-	CityID   *int8 `form:"cityId" binding:"omitempty,min=1"`
-	StateID  *int8 `form:"stateId" binding:"omitempty,min=1"`
+	CityID   *int16 `form:"cityId" binding:"omitempty,min=1"`
+	StateID  *int16 `form:"stateId" binding:"omitempty,min=1"`
 	IsActive *bool `form:"isActive" binding:"omitempty"`
 	MouListQuery
 	Search string `form:"search" binding:"omitempty"`
@@ -120,9 +120,9 @@ type LeadListQuery struct {
 	// StoreMasterID filters by tbl_Leads.StoreMasterID (FK to tbl_StoreMaster.StoreID).
 	StoreMasterID *int64 `form:"storeMasterId" binding:"omitempty,min=1"`
 	// StoreCityID filters by tbl_StoreMaster.CityID (joined on l.StoreMasterID = sm.StoreID).
-	StoreCityID *int8 `form:"storeCityId" binding:"omitempty,min=1"`
+	StoreCityID *int16 `form:"storeCityId" binding:"omitempty,min=1"`
 	// StoreStateID filters by tbl_StoreMaster.StateID (joined on l.StoreMasterID = sm.StoreID).
-	StoreStateID *int8 `form:"storeStateId" binding:"omitempty,min=1"`
+	StoreStateID *int16 `form:"storeStateId" binding:"omitempty,min=1"`
 	// Search matches PatientName, ContactNumber, EmailID, or StoreID (substring, LIKE).
 	Search string `form:"search" binding:"omitempty"`
 	// FitnessStatus filters by tbl_Leads.IsFit (Empty | Not Assessed | On Hold | Fit | UnFit); see domain.ParseLeadListFitnessFilter.
